socket.io redis adapter
